Reorganize tests and add selector escaping tests
Reorganize test structure:
- Create tests/unit/ for component-level tests
- Create tests/integration/ for integration tests
- Simplify smoke.test.js to 3 quick sanity checks

New test files:
- tests/unit/statefulBackend.test.js (3 tests)
- tests/unit/extensionServer.test.js (3 tests)
- tests/unit/oauth.test.js (2 tests)
- tests/unit/unifiedBackend.test.js (9 tests) ✨ NEW
- tests/integration/backend.test.js (3 tests)

Selector escaping tests (unifiedBackend.test.js):
- Verify JSON.stringify properly escapes special characters
- Test single quotes, double quotes, mixed quotes
- Test Unicode (Cyrillic), newlines, backslashes
- Test real-world selectors from bug reports

Total: 23 tests passing (was 11)
New: 12 tests specifically for selector escaping

This prevents regression of the bug where selectors with quotes
(e.g., :has-text('text')) caused JavaScript syntax errors.
